Add form to POST restaurants
Created a form in order to properly create a restaurant. This form also
includes a selection of categories for the 'foodtype' column. This will
allow for a more streamlined way to define the food type and an easier
recall of that information when that feature is to be added.

Note: The categories defined for restaurants need to be organized and
added/removed in a future version. This basic list was created in order
to proceed with the rest of the application and push out a version one
in a timely manner.