Add show for Restaurants resource
The blank restaurants.js file was filled to be able to show all
restaurants and give a button to create one.

Note: This will more than likely be temporary and changed later for a
better place to create a restaurant entry. As of right now this is to
layout the basic content.