Fix compile error on homepage
1. Removed link button in 'Home.js' and kept home blank for now. The
restaurants link was authenticated and mistakenly placed in the home
area.
2. Changed the link in the navbar to show all restaurants, instead of
a nav link to create a restaurant.
3. Fixed an unknown error where the 'g' was missing at the end of
imported apiUrl, which must have been removed by accident.

Note: These link placements are temporary. Once the application is in a
basic usability, the design can be changed to welcome a more user
friendly navigation design.