Fixed the api request for index all restaurants
The API request to pull all restaurants was returning a 401 error.
After checking the backend then re-examining the client, it was
discovered the axios request was not fully written, missing the method
and authorization token. These have been added, and then tested to
ensure the indexing is now working as intended.